requirements.txt
Django==4.2.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-jazzmin==3.0.1
django-ckeditor==6.6.1
django-cors-headers==4.1.0
Pillow==10.0.0
psycopg2-binary>=2.9.9
python-decouple==3.8
whitenoise==6.5.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
stripe==7.4.0
